LPS-54932 Move to the new folder structure
