+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-type-radio/src/com/liferay/dynamic/data/mapping/type/radio/RadioDDMFormFieldRenderer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.type;
+package com.liferay.dynamic.data.mapping.type.radio;
 
 import com.liferay.portal.kernel.template.TemplateConstants;
 import com.liferay.portal.kernel.template.TemplateResource;
