+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-type-radio/test/unit/com/liferay/dynamic/data/mapping/type/radio/RadioDDMFormFieldValueRendererAccessorTest.java
@@ -12,8 +12,10 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.type;
+package com.liferay.dynamic.data.mapping.type.radio;
 
+import com.liferay.dynamic.data.mapping.type.radio.RadioDDMFormFieldValueAccessor;
+import com.liferay.dynamic.data.mapping.type.radio.RadioDDMFormFieldValueRendererAccessor;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portlet.dynamicdatamapping.model.DDMForm;
 import com.liferay.portlet.dynamicdatamapping.model.DDMFormField;
